feat: rename variables
